+++ src/test/java/chess/AutoIncrementTest.java
@@ -1,4 +1,4 @@
-package chess.domain.dao;
+package chess;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
@@ -6,7 +6,6 @@ import java.sql.SQLException;
 import chess.util.JDBCConnector;
 
 public class AutoIncrementTest {
-
 	public static void applyAutoIncrementToZero() throws SQLException {
 		String query = "ALTER TABLE room AUTO_INCREMENT = 0";
 		PreparedStatement pstmt = JDBCConnector.getConnection().prepareStatement(query);
