+++ src/test/java/chess/util/JDBCConnectorTest.java
@@ -1,4 +1,4 @@
-package chess.domain.util;
+package chess.util;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,8 +6,6 @@ import java.sql.Connection;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import chess.util.JDBCConnector;
 
 public class JDBCConnectorTest {
 	@DisplayName("JDBC Connection 테스트")
